+++ studio/src/app/handlers/editor/events/deck/deck-events.handler.tsx
@@ -55,7 +55,9 @@ export class DeckEventsHandler {
             this.el.addEventListener('deckDidChange', this.onDeckChange, false);
             this.el.addEventListener('slideDidChange', this.onSlideChange, false);
             this.el.addEventListener('slideDidLoad', this.onSlideDidLoad, false);
+            this.el.addEventListener('slidesDidLoad', this.onSlidesDidLoad, false);
             this.el.addEventListener('slideDelete', this.onSlideDelete, false);
+            this.el.addEventListener('codeDidChange', this.onCodeChange, false);
 
             this.updateSlideSubscription = this.updateSlideSubject.pipe(debounceTime(500)).subscribe(async (element: HTMLElement) => {
                 await this.updateSlide(element);
@@ -74,7 +76,9 @@ export class DeckEventsHandler {
         this.el.removeEventListener('deckDidChange', this.onDeckChange, true);
         this.el.removeEventListener('slideDidChange', this.onSlideChange, true);
         this.el.removeEventListener('slideDidLoad', this.onSlideDidLoad, true);
+        this.el.removeEventListener('slidesDidLoad', this.onSlidesDidLoad, true);
         this.el.removeEventListener('slideDelete', this.onSlideDelete, true);
+        this.el.removeEventListener('codeDidChange', this.onCodeChange, true);
 
         if (this.updateSlideSubscription) {
             this.updateSlideSubscription.unsubscribe();
@@ -89,8 +93,13 @@ export class DeckEventsHandler {
 
     private onSlideDidLoad = async ($event: CustomEvent) => {
         if ($event && $event.target && $event.target instanceof HTMLElement) {
-            await this.slideToLastSlide($event.target);
             await this.createSlide($event.target);
+        }
+    };
+
+    private onSlidesDidLoad = async ($event: CustomEvent) => {
+        if ($event) {
+            await this.slideToLastSlide();
         }
     };
 
@@ -108,6 +117,22 @@ export class DeckEventsHandler {
         }
 
         this.updateSlideSubject.next($event.detail);
+    };
+
+    private onCodeChange = async ($event: CustomEvent) => {
+        if (!$event || !$event.detail || !($event.detail instanceof HTMLElement)) {
+            return;
+        }
+
+        const element: HTMLElement = $event.detail as HTMLElement;
+
+        const parent: HTMLElement = element.parentElement;
+
+        if (!parent || !parent.nodeName || parent.nodeName.toLowerCase().indexOf('deckgo-slide') <= -1) {
+            return;
+        }
+
+        this.updateSlideSubject.next(parent);
     };
 
     private onSlideInputChange = async ($event: Event) => {
@@ -483,10 +508,10 @@ export class DeckEventsHandler {
                 return;
             }
 
-            // TODO: Should be cleaned on publish: deckgo-untouched|
             let result: string = content.replace(/contenteditable=""|contenteditable="true"|contenteditable/gi, '');
+            result = result.replace(/editable=""|editable="true"|editable/gi, '');
+            result = result.replace(/hydrated/gi, '');
             result = result.replace(/class=""/g, '');
-            result = result.replace(/\s\s+/g, '');
 
             resolve(result);
         });
@@ -500,18 +525,32 @@ export class DeckEventsHandler {
         return SlideTemplate[templateKey];
     }
 
-    private slideToLastSlide(newSlide: HTMLElement): Promise<void> {
+    private slideToLastSlide(): Promise<void> {
         return new Promise<void>(async (resolve) => {
             const deck: HTMLElement = this.el.querySelector('deckgo-deck');
 
-            if (!deck) {
+            if (!deck || !deck.children || deck.children.length <= 0) {
                 resolve();
                 return;
             }
 
-            if (!newSlide.getAttribute('slide_id') && deck.hasChildNodes()) {
-                await (deck as any).slideTo(deck.children && deck.children.length > 0 ? deck.children.length - 1 : 0);
+            const slides: Element[] = Array.from(deck.children).filter((slide: Element) => {
+                return slide.tagName.toLocaleLowerCase().indexOf('deckgo-slide-') > -1
+            });
+
+            if (!slides || slides.length <= 0) {
+                resolve();
+                return;
             }
+
+            const lastSlide: Element = slides[slides.length - 1];
+
+            if (!lastSlide || lastSlide.getAttribute('slide_id')) {
+                resolve();
+                return;
+            }
+
+            await (deck as any).slideTo(slides.length - 1);
 
             resolve();
         });