+++ studio/src/components.d.ts
@@ -8,11 +8,12 @@
 import '@stencil/core';
 
 import '@ionic/core';
-import 'ionicons';
 import 'deckdeckgo';
+import 'deckdeckgo-highlight-code';
 import 'deckdeckgo-inline-editor';
 import 'deckdeckgo-qrcode';
 import 'deckdeckgo-remote';
+import 'ionicons';
 import {
   EventEmitter,
 } from '@stencil/core';
@@ -79,6 +80,7 @@ export namespace Components {
   }
   interface AppEditorToolbarAttributes extends StencilHTMLAttributes {
     'onBlockSlide'?: (event: CustomEvent<boolean>) => void;
+    'onCodeDidChange'?: (event: CustomEvent<HTMLElement>) => void;
     'onDeckDidChange'?: (event: CustomEvent<HTMLElement>) => void;
     'onSlideDelete'?: (event: CustomEvent<HTMLElement>) => void;
     'onSlideDidChange'?: (event: CustomEvent<HTMLElement>) => void;
@@ -179,6 +181,15 @@ export namespace Components {
   interface AppUserMenu {}
   interface AppUserMenuAttributes extends StencilHTMLAttributes {}
 
+  interface AppCode {
+    'codeDidChange': EventEmitter<HTMLElement>;
+    'selectedElement': HTMLElement;
+  }
+  interface AppCodeAttributes extends StencilHTMLAttributes {
+    'codeDidChange'?: EventEmitter<HTMLElement>;
+    'selectedElement'?: HTMLElement;
+  }
+
   interface AppDeckOrSlide {}
   interface AppDeckOrSlideAttributes extends StencilHTMLAttributes {}
 
@@ -227,6 +238,7 @@ declare global {
     'AppSignin': Components.AppSignin;
     'AppEditor': Components.AppEditor;
     'AppUserMenu': Components.AppUserMenu;
+    'AppCode': Components.AppCode;
     'AppDeckOrSlide': Components.AppDeckOrSlide;
     'AppEditorActions': Components.AppEditorActions;
     'AppSlideType': Components.AppSlideType;
@@ -263,6 +275,7 @@ declare global {
     'app-signin': Components.AppSigninAttributes;
     'app-editor': Components.AppEditorAttributes;
     'app-user-menu': Components.AppUserMenuAttributes;
+    'app-code': Components.AppCodeAttributes;
     'app-deck-or-slide': Components.AppDeckOrSlideAttributes;
     'app-editor-actions': Components.AppEditorActionsAttributes;
     'app-slide-type': Components.AppSlideTypeAttributes;
@@ -444,6 +457,12 @@ declare global {
     new (): HTMLAppUserMenuElement;
   };
 
+  interface HTMLAppCodeElement extends Components.AppCode, HTMLStencilElement {}
+  var HTMLAppCodeElement: {
+    prototype: HTMLAppCodeElement;
+    new (): HTMLAppCodeElement;
+  };
+
   interface HTMLAppDeckOrSlideElement extends Components.AppDeckOrSlide, HTMLStencilElement {}
   var HTMLAppDeckOrSlideElement: {
     prototype: HTMLAppDeckOrSlideElement;
@@ -498,6 +517,7 @@ declare global {
     'app-signin': HTMLAppSigninElement
     'app-editor': HTMLAppEditorElement
     'app-user-menu': HTMLAppUserMenuElement
+    'app-code': HTMLAppCodeElement
     'app-deck-or-slide': HTMLAppDeckOrSlideElement
     'app-editor-actions': HTMLAppEditorActionsElement
     'app-slide-type': HTMLAppSlideTypeElement
@@ -534,6 +554,7 @@ declare global {
     'app-signin': HTMLAppSigninElement;
     'app-editor': HTMLAppEditorElement;
     'app-user-menu': HTMLAppUserMenuElement;
+    'app-code': HTMLAppCodeElement;
     'app-deck-or-slide': HTMLAppDeckOrSlideElement;
     'app-editor-actions': HTMLAppEditorActionsElement;
     'app-slide-type': HTMLAppSlideTypeElement;