+++ webcomponents/inline-editor/src/components.d.ts
@@ -14,14 +14,14 @@ export namespace Components {
 
   interface DeckgoInlineEditor {
     'attachTo': HTMLElement;
-    'containerAttribute': string;
+    'containers': string;
     'mobile': boolean;
     'stickyDesktop': boolean;
     'stickyMobile': boolean;
   }
   interface DeckgoInlineEditorAttributes extends StencilHTMLAttributes {
     'attachTo'?: HTMLElement;
-    'containerAttribute'?: string;
+    'containers'?: string;
     'mobile'?: boolean;
     'stickyDesktop'?: boolean;
     'stickyMobile'?: boolean;