+++ CHANGELOG.md
@@ -1,3 +1,33 @@
+<a name="1.0.0-beta.37"></a>
+# [1.0.0-beta.37](https://github.com/deckgo/deckdeckgo/compare/v1.0.0-beta.36...v1.0.0-beta.37) (2019-04-30)
+
+### Web Components
+
+* highlight-code: v1.0.0-alpha.14 ([CHANGELOG](https://github.com/deckgo/deckdeckgo/blob/master/webcomponents/highlight-code/CHANGELOG.md))
+
+### Applications
+
+* docs: v1.0.0-alpha.7 ([CHANGELOG](https://github.com/deckgo/deckdeckgo/blob/master/docs/CHANGELOG.md))
+
+<a name="1.0.0-beta.36"></a>
+# [1.0.0-beta.36](https://github.com/deckgo/deckdeckgo/compare/v1.0.0-beta.35...v1.0.0-beta.36) (2019-04-30)
+
+### Web Components
+
+* core: v1.0.0-beta.32 ([CHANGELOG](https://github.com/deckgo/deckdeckgo/blob/master/webcomponents/core/CHANGELOG.md))
+
+<a name="1.0.0-beta.35"></a>
+# [1.0.0-beta.35](https://github.com/deckgo/deckdeckgo/compare/v1.0.0-beta.34...v1.0.0-beta.35) (2019-04-30)
+
+### Web Components
+
+* highlight-code: v1.0.0-alpha.13 ([CHANGELOG](https://github.com/deckgo/deckdeckgo/blob/master/webcomponents/highlight-code/CHANGELOG.md))
+* inline-editor: v1.0.0-alpha.5 ([CHANGELOG](https://github.com/deckgo/deckdeckgo/blob/master/webcomponents/inline-editor/CHANGELOG.md))
+
+### Applications
+
+* docs: v1.0.0-alpha.6 ([CHANGELOG](https://github.com/deckgo/deckdeckgo/blob/master/docs/CHANGELOG.md))
+
 <a name="1.0.0-beta.34"></a>
 # [1.0.0-beta.34](https://github.com/deckgo/deckdeckgo/compare/v1.0.0-beta.33...v1.0.0-beta.34) (2019-04-19)