+++ docs/CHANGELOG.md
@@ -1,3 +1,17 @@
+<a name="1.0.0-alpha.7"></a>
+# 1.0.0-alpha.7 (2019-04-30)
+
+### Features
+
+* this highlight-code component has now a default color for the background of the highlighted lines
+
+<a name="1.0.0-alpha.6"></a>
+# 1.0.0-alpha.6 (2019-04-30)
+
+### Features
+
+* the highlight-code component could now be optionally set as `editable`
+
 <a name="1.0.0-alpha.5"></a>
 # 1.0.0-alpha.5 (2019-04-18)