+++ webcomponents/inline-editor/CHANGELOG.md
@@ -1,3 +1,10 @@
+<a name="1.0.0-alpha.5"></a>
+# 1.0.0-alpha.5 (2019-04-30)
+
+### Breaking change
+
+* modify the default `containers` value from `slot` to `h1,h2,h3,h4,h5,h6,div`
+
 <a name="1.0.0-alpha.4"></a>
 # 1.0.0-alpha.4 (2019-04-18)